requirements.txt
python-telegram-bot[job-queue]==21.10
flask==3.1.0
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
python-dotenv==1.0.0
waitress==2.1.2
gunicorn==23.0.0
